Training/requirements.txt
langchain>=0.2.5
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=0.0.10
chromadb>=0.5.0
tiktoken>=0.7.0
pypdf>=4
python-dotenv>=1.0.1
langgraph>=0.1.0
ipython>=8.0.0
jupyter>=1.0.0
